+++ docs/pages/installation/ConstructorOptions.vue
@@ -21,7 +21,7 @@
                 api,
                 usage: `
                 Vue.use(Buefy, {
-                    defaultIconPack: 'fa',
+                    defaultIconPack: 'fas',
                     defaultContainerElement: '#content',
                     // ...
                 })`