+++ docs/pages/installation/Start.vue
@@ -81,7 +81,8 @@
         <h3 class="title">Include Material Design Icons</h3>
         <b-message type="is-info">
             By default Buefy uses <a href="https://materialdesignicons.com/" target="_blank">Material Design Icons</a>,
-            if you want to swap to <a href="http://fontawesome.io/" target="_blank">FontAwesome</a>, go to
+            if you want to swap to <a href="http://fontawesome.io/" target="_blank">FontAwesome 4</a> 
+            or <a href="https://fontawesome.com/" target="_blank">FontAwesome 5</a>, go to
             <router-link to="/documentation/constructor-options">constructor options</router-link>.
         </b-message>
         <b-message type="is-warning">